all class definitions, constants, and typdef merge into one
